그 외에 와이어 프레임에 따라 네비게이션바를 수정하고, 검색바를 따로 분리해두었습니다.
